se corrige MRE

+++ Capa-Datos/modeloCadena/src/main/java/com/example/entidades/IngredientePlato.java
@@ -28,5 +28,5 @@ public class IngredientePlato implements Serializable {
     private Ingrediente ingrediente;
 
     @Column(name = "cantidad")
-    private double cantidad;
+    private Integer cantidad;
 }
