+++ Capa-Datos/modeloCadena/src/main/java/com/example/entidades/Plato.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 
 @AllArgsConstructor
 @NoArgsConstructor
@@ -25,7 +26,7 @@ public class Plato implements Serializable {
     private String descripcion;
 
     @Column(name = "precio")
-    private double precio;
+    private BigInteger precio;
 
     @ManyToOne
     @JoinColumn(name = "Menuid")
